Negotiate feature bits for virtio-net
