+++ kernel/src/arch_rv32/drivers/mod.rs
@@ -1,0 +1,1 @@
+pub mod virtio_net;
